Added bulgarian-phonetic-simple keymap.
The bulgarian-phonetic keymap has a lot of special character support.
This means a lot of key presses do not generate the final character
immediately, but wait for a while to see if there won't be other
characters to form a special character. In case you never use those
characters this is a downside. This simplified keymap has just the
standard letter keys mapped, so all characters appear immediately after
they've been pressed.